api: switch to generating openapi spec directly instead of using the endpoint
The endpoint seems to be not available anymore in api platform 3.2, or does not support
json at least.
And we can switch to yaml like this, which is easier to merge (and read).

This requires to mark some services public that we need the DI container
to wire for the test.