api: embed preferredContentTypes in the category when requested as subresource
This leads currently to requests to
api/content_types?categories=%2Fapi%2Fcategories%2F{id}%2F
which we cannot cache yet.
They are executed when you select a prototype while creating a camp.
We also cannot prefetch all collection urls for the preferredContentTypes.
The request somehow takes a while.
In a camp there are not many categories, and the amount of contentTypes
that would be joined is also limited.
Currently, preferredContentTypes is anyway mostly empty.

Issue: #6461